ml_back/requirements.txt
flask
flask-cors
opencv-python
numpy
pillow
ultralytics
